se agregaron y corrigieron todas las funciones de la view busqueda

+++ src/main/java/dao/ReservaDao.java
@@ -6,6 +6,7 @@ import javax.persistence.EntityManager;
 import javax.persistence.EntityTransaction;
 import javax.persistence.TypedQuery;
 
+import modelos.Huesped;
 import modelos.Reserva;
 import javax.persistence.Query;
 
@@ -27,16 +28,37 @@ public class ReservaDao {
 
 	}
 
-	public void eliminar(Reserva reserva) {
-		EntityTransaction tx = em.getTransaction();
-		tx.begin();
-		em.remove(reserva);
-		tx.commit();
-	}
-
 	public List<Reserva> obtenerTodos() {
 		TypedQuery<Reserva> query = em.createQuery("SELECT h FROM Reserva h", Reserva.class);
 		return query.getResultList();
+	}
+	
+	public List<Reserva> listarPorId(int id) {
+	    TypedQuery<Reserva> query = em.createQuery("SELECT r FROM Reserva r WHERE r.id = :id", Reserva.class);
+	    query.setParameter("id", (long) id);
+	    List<Reserva> resultados = query.getResultList();
+	    if (resultados.isEmpty()) {
+	        resultados = obtenerTodos(); 
+	    }
+	    return resultados;
+	}
+
+	
+	public void eliminarReserva(Long id) {
+	    EntityTransaction tx = em.getTransaction();
+	    try {
+	        tx.begin();
+	        Reserva reserva = em.find(Reserva.class, id);
+	        if (reserva != null) {
+	            em.remove(reserva);
+	        }
+	        tx.commit();
+	    } catch (Exception e) {
+	        if (tx != null && tx.isActive()) {
+	            tx.rollback();
+	        }
+	        throw e;
+	    }
 	}
 
 	public void actualizarReserva(Reserva reserva) {
